Seperate blacklist

+++ hippiestation/code/modules/reagents/chemistry/reagents.dm
@@ -79,31 +79,32 @@
 		if(src.reagent_state == LIQUID) //LIQUID
 			if(atom && istype(atom, /obj/effect/particle_effect))
 				volume = volume * LIQUID_PARTICLE_EFFECT_EFFICIENCY//big nerf to smoke and foam duping
-
-			for(var/obj/effect/decal/cleanable/chempile/c in T.contents)//handles merging existing chempiles
-				if(c.reagents)
-					if(touch_msg)
-						c.add_fingerprint(touch_mob)
-					c.reagents.add_reagent("[src.id]", volume)
-					var/mixcolor = mix_color_from_reagents(c.reagents.reagent_list)
-					c.add_atom_colour(mixcolor, FIXED_COLOUR_PRIORITY)
-					if(c.reagents.total_volume >= 10)
-						var/obj/effect/liquid/W = new /obj/effect/liquid(T)
-						c.reagents.trans_to(W, c.reagents.total_volume)
-						qdel(c)
-						W.depth = CLAMP(W.depth + (c.reagents.total_volume / REAGENT_TO_DEPTH), 0, MAX_INITIAL_DEPTH)
-						W.update_depth()
+			var/can_be_chempile = !is_type_in_typecache(src, GLOB.chempile_reagent_blacklist)
+			if(can_be_chempile)
+				for(var/obj/effect/decal/cleanable/chempile/c in T.contents)//handles merging existing chempiles
+					if(c.reagents)
+						if(touch_msg)
+							c.add_fingerprint(touch_mob)
+						c.reagents.add_reagent("[src.id]", volume)
+						var/mixcolor = mix_color_from_reagents(c.reagents.reagent_list)
+						c.add_atom_colour(mixcolor, FIXED_COLOUR_PRIORITY)
+						if(c.reagents.total_volume >= 10)
+							var/obj/effect/liquid/W = new /obj/effect/liquid(T)
+							c.reagents.trans_to(W, c.reagents.total_volume)
+							qdel(c)
+							W.depth = CLAMP(W.depth + (c.reagents.total_volume / REAGENT_TO_DEPTH), 0, MAX_INITIAL_DEPTH)
+							W.update_depth()
+							return TRUE
+						if(c.reagents && c.reagents.total_volume < 5 & NO_REACT)
+							DISABLE_BITFIELD(c.reagents.flags, NO_REACT)
 						return TRUE
-					if(c.reagents && c.reagents.total_volume < 5 & NO_REACT)
-						DISABLE_BITFIELD(c.reagents.flags, NO_REACT)
-					return TRUE
 			for(var/obj/effect/liquid/L in T.contents)//handles merging existing liquids
 				if(L.reagents)
 					L.reagents.add_reagent("[src.id]", volume * 4)
 					L.depth = CLAMP(L.depth + (volume / REAGENT_TO_DEPTH), 0, MAX_INITIAL_DEPTH)
 					L.update_depth()
 					return TRUE
-			if(volume < 10)
+			if(volume < 10 && !can_be_chempile)
 				var/obj/effect/decal/cleanable/chempile/C = new /obj/effect/decal/cleanable/chempile(T)
 				if(C.reagents)
 					if(touch_msg)
